+++ packages/viridis/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.4.0
+
+- Remove `Color.from()` (use `Color.hex()` instead)
+- Support more gradient types in `Gradient.toString()`
+- Update custom gradient example
+
 ## 1.3.1
 
 - Re-generate builtin color palettes in main README file